+++ app/Http/Controllers/ProductController.php
@@ -5,25 +5,44 @@ namespace App\Http\Controllers;
 use App\Http\Requests\Product\CreateProductRequest;
 use App\Http\Requests\Product\EditProductRequest;
 use App\Models\Product;
+use Illuminate\Contracts\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class ProductController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index(string $id = null)
+    public function index(Request $request, string $id = null)
     {
-        if(!$id)
-    {
-        $products = Product::with(['category:id,title', 'brand'])->orderby('id', 'desc')->paginate(2);
-        return response()->json($products);
-    }
-    else
-    {
-        $product = Product::find($id);
-        return response()->json($product);
-    }
+        if($request->user()->can('read.product'))
+        {
+            if(!$id)
+            {
+                $products = new Product();
+                $products = $products->with(['category:id,title', 'brand', 'warranties', 'labels']);
+                if($request->most_sold)
+                {
+                    $products = $products->withCount('orders');
+                    $products = $products->orderBy('orders_count', 'desc');
+                    $products = $products->skip(0)->take(3);
+                    $products = $products->get();
+                    return response()->json($products);
+                }
+                $products = $products->orderby('id', 'desc')->paginate(5);
+                return response()->json($products);
+            }
+            else
+            {
+                $product = Product::with(['category:id,title', 'brand', 'warranties', 'labels'])->find($id);
+                return response()->json($product);
+            }
+        }
+        else
+        {
+            return response()->json('User does not have the permission', 403);
+        }
     }
 
     /**
@@ -31,8 +50,18 @@ class ProductController extends Controller
      */
     public function store(CreateProductRequest $request)
     {
-        $product = Product::create($request->toArray());
-        return response()->json($product);
+        if($request->user()->can('create.product'))
+        {
+            $path = $request->file('image_path')->store('public/products');
+            $product = Product::create($request->merge(["image_path" => $path])->toArray());
+            $product->warranties()->attach($request->warranty_ids);
+            $product->labels()->attach($request->label_ids);
+            return response()->json($product);
+        }
+        else
+        {
+            return response()->json('User does not have the permission', 403);
+        }
     }
 
     /**
@@ -49,17 +78,31 @@ class ProductController extends Controller
      */
     public function update(EditProductRequest $request, string $id)
     {
-        $product = Product::find($id)->update($request->toArray());
-        return response()->json($product);
+        if($request->user()->can('update.product'))
+        {
+            $product = Product::find($id)->update($request->toArray());
+            return response()->json($product);
+        }
+        else
+        {
+            return response()->json('User does not have the permission', 403);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
-        $product = Product::destroy($id);
-        return response()->json($product);
+        if($request->user()->can('delete.product'))
+        {
+            $product = Product::destroy($id);
+            return response()->json($product);
+        }
+        else
+        {
+            return response()->json('User does not have the permission', 403);
+        }
     }
 }
